feat(db): add cover_url column to tags table
This change adds a new `cover_url` column to the `tags` table in the database schema. This allows storing a cover image URL for each tag, which can be used to display a visual representation of the tag on the frontend.

The seed script has also been updated to generate random cover URLs for the tags being inserted, using a predefined list of Unsplash image URLs.