feat(bookshelves): add user information to bookshelf card
Adds a new component `BookShelveUserInfo` to display the user information
associated with a bookshelf. This includes the user's avatar, name, and the
time since the bookshelf was created.

The `BookShelveCard` component has been updated to include this new user
information section, which is wrapped in a `Suspense` component to provide a
loading state while the user data is being fetched.