feat(authors): add layout option to AuthorList component
This change adds a new `layout` option to the `AuthorList` component, allowing the
layout of the author cards to be customized. The new layouts include 'default',
'compact', 'narrow', and 'single'. This provides more flexibility in how the
author list is displayed, depending on the specific use case.

The changes also include updating the `AdminAuthorsPage` component to use the
'narrow' layout for the author list, which is more suitable for the admin
dashboard context.